backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-genai==0.8.0
pydantic==2.9.2
